Make it an installable library
