+++ src/didactic_datamining/didactic_kmeans.py
@@ -1,4 +1,3 @@
-import json
 import itertools
 import numpy as np
 import matplotlib.pyplot as plt
